Seperated IOHandler and PatientList out into there own classes. They are both singletons.
