+++ app/src/main/java/luckyseven/clinicapplication/FormerClinicApplication.java
@@ -1,5 +1,6 @@
 package groupproject;
 
+import luckyseven.clinicapplication.*;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
